+++ src/Handlers/AbstractHandler.php
@@ -1,14 +1,11 @@
 <?php
 /**
  * Copyright (c) 2012 ScientiaMobile, Inc.
- *
  * This program is free software: you can redistribute it and/or modify
  * it under the terms of the GNU Affero General Public License as
  * published by the Free Software Foundation, either version 3 of the
  * License, or (at your option) any later version.
- *
  * Refer to the COPYING.txt file distributed with this package.
- *
  *
  * @category   WURFL
  * @package    WURFL
@@ -83,7 +80,6 @@ abstract class AbstractHandler
      * sets the logger
      *
      * @var \Psr\Log\LoggerInterface $logger
-     *
      * @return \Wurfl\Handlers\AbstractHandler
      */
     public function setLogger(LoggerInterface $logger = null)
@@ -97,7 +93,6 @@ abstract class AbstractHandler
      * sets the Persitence Cache
      *
      * @var \Wurfl\Storage\Storage $persistenceProvider
-     *
      * @return \Wurfl\Handlers\AbstractHandler
      */
     public function setPersistenceProvider(Storage $persistenceProvider)
@@ -409,23 +404,78 @@ abstract class AbstractHandler
             return Constants::GENERIC_WEB_BROWSER;
         }
 
-        $mobile  = Utils::isMobileBrowser($userAgent);
-        $desktop = Utils::isDesktopBrowser($userAgent);
-
-        if (!$desktop) {
-            $deviceId = Utils::getMobileCatchAllId($userAgent);
-
-            if ($deviceId !== Constants::NO_MATCH) {
-                return $deviceId;
-            }
+        if (Utils::checkIfContains($userAgent, 'CoreMedia')) {
+            return 'apple_iphone_coremedia_ver1';
+        }
+        if (Utils::checkIfContains($userAgent, 'Windows CE')) {
+            return 'generic_ms_mobile';
+        }
+        if (Utils::checkIfContains($userAgent, 'UP.Browser/7.2')) {
+            return 'opwv_v72_generic';
+        }
+        if (Utils::checkIfContains($userAgent, 'UP.Browser/7')) {
+            return 'opwv_v7_generic';
+        }
+        if (Utils::checkIfContains($userAgent, 'UP.Browser/6.2')) {
+            return 'opwv_v62_generic';
+        }
+        if (Utils::checkIfContains($userAgent, 'UP.Browser/6')) {
+            return 'opwv_v6_generic';
+        }
+        if (Utils::checkIfContains($userAgent, 'UP.Browser/5')) {
+            return 'upgui_generic';
+        }
+        if (Utils::checkIfContains($userAgent, 'UP.Browser/4')) {
+            return 'uptext_generic';
+        }
+        if (Utils::checkIfContains($userAgent, 'UP.Browser/3')) {
+            return 'uptext_generic';
+        }
+        // Series 60
+        if (Utils::checkIfContains($userAgent, 'Series60')) {
+            return 'nokia_generic_series60';
+        }
+        // Access/Net Front
+        if (Utils::checkIfContainsAnyOf($userAgent, array('NetFront/3.0', 'ACS-NF/3.0'))) {
+            return 'generic_netfront_ver3';
+        }
+        if (Utils::checkIfContainsAnyOf($userAgent, array('NetFront/3.1', 'ACS-NF/3.1'))) {
+            return 'generic_netfront_ver3_1';
+        }
+        if (Utils::checkIfContainsAnyOf($userAgent, array('NetFront/3.2', 'ACS-NF/3.2'))) {
+            return 'generic_netfront_ver3_2';
+        }
+        if (Utils::checkIfContainsAnyOf($userAgent, array('NetFront/3.3', 'ACS-NF/3.3'))) {
+            return 'generic_netfront_ver3_3';
+        }
+        if (Utils::checkIfContains($userAgent, 'NetFront/3.4')) {
+            return 'generic_netfront_ver3_4';
+        }
+        if (Utils::checkIfContains($userAgent, 'NetFront/3.5')) {
+            return 'generic_netfront_ver3_5';
+        }
+        if (Utils::checkIfContains($userAgent, 'NetFront/4.0')) {
+            return 'generic_netfront_ver4_0';
+        }
+        // Contains Mozilla/, but not at the beginning of the UA
+        // ie: MOTORAZR V8/R601_G_80.41.17R Mozilla/4.0 (compatible; MSIE 6.0 Linux; MOTORAZR V88.50) Profile/MIDP-2.0 Configuration/CLDC-1.1 Opera 8.50[zh]
+        if (strpos($userAgent, 'Mozilla/') > 0) {
+            return Constants::GENERIC_XHTML;
+        }
+        if (Utils::checkIfContainsAnyOf(
+            $userAgent,
+            array('Obigo', 'AU-MIC/2', 'AU-MIC-', 'AU-OBIGO/', 'Teleca Q03B1')
+        )
+        ) {
+            return Constants::GENERIC_XHTML;
+        }
+        // DoCoMo
+        if (Utils::checkIfStartsWithAnyOf($userAgent, array('DoCoMo', 'KDDI'))) {
+            return 'docomo_generic_jap_ver1';
         }
 
-        if ($mobile) {
+        if (Utils::isMobileBrowser($userAgent)) {
             return Constants::GENERIC_MOBILE;
-        }
-
-        if ($desktop) {
-            return Constants::GENERIC_WEB_BROWSER;
         }
 
         return Constants::GENERIC;