+++ src/VirtualCapability/Tool/Device.php
@@ -78,17 +78,18 @@ class Device
      * @var array
      */
     protected static $windowsMap = array(
-        '3.1' => 'NT 3.1',
-        '3.5' => 'NT 3.5',
-        '4.0' => 'NT 4.0',
-        '5.0' => '2000',
-        '5.1' => 'XP',
-        '5.2' => 'XP',
-        '6.0' => 'Vista',
-        '6.1' => '7',
-        '6.2' => '8',
-        '6.3' => '8.1',
-        '6.4' => '10',
+        '3.1'  => 'NT 3.1',
+        '3.5'  => 'NT 3.5',
+        '4.0'  => 'NT 4.0',
+        '5.0'  => '2000',
+        '5.1'  => 'XP',
+        '5.2'  => 'XP',
+        '6.0'  => 'Vista',
+        '6.1'  => '7',
+        '6.2'  => '8',
+        '6.3'  => '8.1',
+        '6.4'  => '10',
+        '10.0' => '10',
     );
 
     /**
@@ -109,7 +110,7 @@ class Device
         if (strpos($this->device_ua, 'Windows') !== false) {
             $matches = array();
 
-            if (preg_match('/Windows NT ([0-9]\.[0-9])/', $this->os->name, $matches)) {
+            if (preg_match('/Windows NT ([0-9]+?\.[0-9])/', $this->os->name, $matches)) {
                 $this->os->name    = 'Windows';
                 $this->os->version = array_key_exists($matches[1], self::$windowsMap) ? self::$windowsMap[$matches[1]]
                     : $matches[1];