+++ src/VirtualCapability/UserAgentTool.php
@@ -51,26 +51,28 @@ class UserAgentTool
      */
     protected function assignProperties(Tool\Device $device)
     {
-        //Is UA Windows Mobile? - WP before Android
+        //Is UA Windows Mobile?
         if ($device->os->setContains($device->device_ua, 'Windows CE', 'Windows Mobile')) {
             $device->browser->set('IE Mobile');
 
             return $device;
         }
 
+        //Is UA Windows Phone OS? - WP before Android
         if (strpos($device->device_ua, 'Windows Phone') !== false) {
-            // Is UA Windows Phone OS?
             if ($device->os->setRegex(
                 $device->device_ua,
-                '/Windows Phone(?: OS)? ([0-9]\.[0-9])/',
+                '/Windows Phone(?: OS)? ([0-9]+\.[0-9])/',
                 'Windows Phone',
                 1
             )
             ) {
-                $device->browser->set('IE Mobile');
-                $device->browser->setRegex($device->device_ua, '/IEMobile\/(\d+\.\d+)/', 'IE Mobile', 1);
-
-                return $device;
+                if ($device->browser->setRegex($device->browser_ua, '/UCBrowser\/([0-9]+)\./', 'UC Browser', 1)) {
+                    return $device;
+                }
+                if ($device->browser->setRegex($device->browser_ua, '/IEMobile\/(\d+\.\d+)/', 'IE Mobile', 1)) {
+                    return $device;
+                }
             }
         }
 
@@ -95,6 +97,22 @@ class UserAgentTool
                 $device->os->version
             )
             ) {
+                return $device;
+            }
+
+            //Is UA Opera?
+            if ($device->browser->setRegex($device->browser_ua, '/OPR\/([0-9]?[0-9])\.?/', 'Opera', 1)) {
+                return $device;
+            }
+
+            //Is 360Browser?
+            if (strpos($device->browser_ua, 'Aphone Browser') !== false || strpos(
+                    $device->browser_ua,
+                    '360browser'
+                ) !== false
+            ) {
+                $device->browser->set('360 Browser', null);
+
                 return $device;
             }
 
@@ -201,7 +219,18 @@ class UserAgentTool
             if ($device->browser->setRegex(
                 $device->browser_ua,
                 '/^Mozilla\/[45]\.0.+?like Mac OS X.+?CriOS\/([0-9]+?)\.[0-9].+?Mobile\/[0-9A-Za-z]+ Safari\/[0-9A-Za-z]+\./',
-                'Chrome Mobile iOS',
+                'Chrome Mobile on iOS',
+                1
+            )
+            ) {
+                return $device;
+            }
+
+            //Is UA Opera Mini on iOS?
+            if ($device->browser->setRegex(
+                $device->browser_ua,
+                '/^Mozilla\/[45]\.0.+?like Mac OS X.+?OPiOS\/([0-9]+?)\.[0-9].+?Mobile\/[0-9A-Za-z]+ Safari\/[0-9A-Za-z]+\./',
+                'Opera Mini on iOS',
                 1
             )
             ) {
@@ -474,9 +503,23 @@ class UserAgentTool
 
         // Desktop Browsers
 
-        //MSIE
-        if (strpos($device->device_ua, 'Trident') !== false || strpos($device->device_ua, 'MSIE') !== false) {
-            //MSIE 10 and below
+        //360 Browser
+        if ((strpos($device->device_ua, '360Browser') !== false || strpos(
+                    $device->device_ua,
+                    ' 360SE'
+                ) !== false) && $device->os->setRegex(
+                $device->device_ua,
+                '/^Mozilla\/[0-9]\.0 .+?((?:Windows|Linux|PPC|Intel) [a-zA-Z0-9 _\.\-]+).+(?:360Browser|360SE)/',
+                1
+            )
+        ) {
+            $device->browser->set('360 Browser', null);
+
+            return $device;
+        }
+
+        //MSIE - If UA says MSIE
+        if (strpos($device->device_ua, 'MSIE') !== false) {
             if ($device->os->setRegex(
                 $device->device_ua,
                 '/^Mozilla\/[0-9]\.0 \(compatible; MSIE ([0-9][0-9]?\.[0-9][0-9]?); ((?:Windows NT [0-9]\.[0-9])|(?:Windows [0-9]\.[0-9])|(?:Windows [0-9]+)|(?:Mac_PowerPC))/',
@@ -486,8 +529,13 @@ class UserAgentTool
                 $device->browser->set('IE', $device->os->getLastRegexMatch(1));
 
                 return $device;
-            } //MSIE 11 and above
-            else if ($device->os->setRegex(
+            }
+        }
+
+        //MSIE - If UA says Trident
+        if (strpos($device->device_ua, 'Trident') !== false) {
+            //MSIE 11 does not say MSIE and needs this
+            if ($device->os->setRegex(
                 $device->device_ua,
                 '#^Mozilla/[45]\.0 \((Windows NT [0-9]\.[0-9]);.+Trident.+; rv:([0-9]+)\.[0-9]+#',
                 1
@@ -508,6 +556,23 @@ class UserAgentTool
         ) {
             $device->browser->set('Yandex browser', $device->os->getLastRegexMatch(2));
 
+            return $device;
+        }
+
+        //Opera - OPR
+        if (strpos($device->device_ua, 'OPR') !== false
+            && $device->os->setRegex($device->device_ua, '/^Mozilla\/[0-9]\.0 .+?((?:Windows|Linux|PPC|Intel) [a-zA-Z0-9 _\.\-]+).+Chrome\/.+OPR\/([0-9]+?)\./', 1)
+        ) {
+            $device->browser->set('Opera', $device->os->getLastRegexMatch(2));
+            return $device;
+        }
+
+        //Opera - Old UA
+        if (strpos($device->device_ua, 'Opera') !== false
+            && $device->os->setRegex($device->device_ua, '/^Opera\/([0-9]?[0-9]\.[0-9][0-9]?) .+?((?:Windows|Linux|PPC|Intel) [a-zA-Z0-9 _\.\-]+) ?;/', 2)
+        ) {
+            $device->browser->set('Opera', $device->os->getLastRegexMatch(1));
+            $device->browser->setRegex($device->browser_ua, '/^Opera\/.+? Version\/([0-9]?[0-9]\.[0-9][0-9]?)/', null, 1);
             return $device;
         }
 
@@ -573,24 +638,6 @@ class UserAgentTool
 
                 return $device;
             }
-        }
-
-        //Opera
-        if (strpos($device->device_ua, 'Opera') !== false && $device->os->setRegex(
-                $device->device_ua,
-                '/^Opera\/([0-9]?[0-9]\.[0-9][0-9]?) .+?((?:Windows|Linux|PPC|Intel) [a-zA-Z0-9 _\.\-]+) ?;/',
-                2
-            )
-        ) {
-            $device->browser->set('Opera', $device->os->getLastRegexMatch(1));
-            $device->browser->setRegex(
-                $device->browser_ua,
-                '/^Opera\/.+? Version\/([0-9]?[0-9]\.[0-9][0-9]?)/',
-                null,
-                1
-            );
-
-            return $device;
         }
 
         return $device;